(Tests + Post-processing) First working version for output. Should have in config an output configuration for keys and stuff :)
