+++ ppa_splitter/cli.py
@@ -83,6 +83,7 @@ def dispatch(
             if value:
                 print("\t{} tokens in {} dataset".format(value, key))
 
+
     if concat:
         print("==============")
         print("Concatenating")
